backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx>=0.26,<0.28
supabase==2.10.0
python-dotenv==1.0.1
PyJWT[crypto]==2.10.1
slowapi==0.1.9
